xrpl-bot-backend/requirements.txt
anyio==4.9.0
base58==2.1.1
blinker==1.6.2
certifi==2025.6.15
click==8.1.3
Deprecated==1.2.18
ECPy==1.2.5
exceptiongroup==1.3.0
Flask==2.3.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==25.0
pycryptodome==3.23.0
pygame==2.5.0
python-telegram-bot==22.1
sniffio==1.3.1
types-Deprecated==1.2.15.20250304
typing_extensions==4.14.0
websockets==15.0.1
Werkzeug==2.3.6
wrapt==1.17.2
xrpl-py==4.2.0
